Slim GUI examples
