+++ examples/egui-skia/Cargo.toml
@@ -8,7 +8,7 @@ name = "egui-skia"
 path = "main.rs"
 
 [dependencies]
-gosub_engine = { path = "../../crates/gosub_engine", features = ["sqlite_cookie_store"] }
+gosub_engine = { path = "../../crates/gosub_engine" }
 gosub_render_pipeline = { path = "../../crates/gosub_render_pipeline" }
 gosub_renderer_skia = { path = "../../crates/gosub_renderer_skia" }
 eframe = { workspace = true, features = ["wgpu"] }
