+++ examples/gtk4-cairo/Cargo.toml
@@ -8,7 +8,7 @@ name = "gtk4-cairo"
 path = "main.rs"
 
 [dependencies]
-gosub_engine = { path = "../../crates/gosub_engine", features = ["sqlite_cookie_store"] }
+gosub_engine = { path = "../../crates/gosub_engine" }
 gosub_config = { path = "../../crates/gosub_config" }
 gosub_render_pipeline = { path = "../../crates/gosub_render_pipeline" }
 gosub_renderer_cairo = { path = "../../crates/gosub_renderer_cairo" }
