+++ examples/gtk4-skia-gpu/main.rs
@@ -12,7 +12,6 @@
 #[link(name = "GL")]
 extern "C" {}
 
-use gosub_engine::cookies::SqliteCookieStore;
 use gosub_engine::events::{EngineEvent, NavigationEvent, TabCommand};
 use gosub_engine::storage::{InMemorySessionStore, PartitionPolicy, SqliteLocalStore, StorageService};
 use gosub_engine::tab::{TabDefaults, TabId};
@@ -69,122 +68,6 @@ fn get_bound_fbo() -> u32 {
 
 // ── Application ───────────────────────────────────────────────────────────────
 
-/// Map the engine's cursor kind to a named GTK cursor on the page widget.
-fn apply_cursor(area: &GLArea, cursor: gosub_engine::events::CursorShape) {
-    use gosub_engine::events::CursorShape;
-    let name = match cursor {
-        CursorShape::Default => "default",
-        CursorShape::Pointer => "pointer",
-        CursorShape::Text => "text",
-        CursorShape::Resize => "nwse-resize",
-    };
-    area.set_cursor_from_name(Some(name));
-}
-
-/// Ctrl+C/X/V inside the page: the engine only sees keys, the clipboard is the embedder's. Copy/cut
-/// arrive as text to store; a paste request is answered by reading the clipboard and sending it
-/// back as `TextInput`.
-fn handle_clipboard_event(evt: &EngineEvent, area: &GLArea, tab: &Rc<RefCell<gosub_engine::tab::TabHandle>>) -> bool {
-    let clipboard = area.display().clipboard();
-    match evt {
-        EngineEvent::ClipboardWrite { text, .. } => {
-            clipboard.set_text(text);
-            true
-        }
-        EngineEvent::PasteRequested { .. } => {
-            let tab = tab.borrow().clone();
-            clipboard.read_text_async(gtk4::gio::Cancellable::NONE, move |res| {
-                let Ok(Some(text)) = res else {
-                    return;
-                };
-                let text = text.to_string();
-                TOKIO_RT.spawn(async move {
-                    let _ = tab.send(TabCommand::TextInput { text }).await;
-                });
-            });
-            true
-        }
-        _ => false,
-    }
-}
-
-/// GDK key names mostly match DOM `KeyboardEvent.key`; translate the ones that don't. X11 gives
-/// Shift+Tab its own `ISO_Left_Tab` keysym.
-fn key_down_command(key: gtk4::gdk::Key, state: gtk4::gdk::ModifierType) -> Option<TabCommand> {
-    use gosub_engine::events::Modifiers;
-    let mut modifiers = Modifiers::empty();
-    if state.contains(gtk4::gdk::ModifierType::SHIFT_MASK) {
-        modifiers |= Modifiers::SHIFT;
-    }
-    if state.contains(gtk4::gdk::ModifierType::CONTROL_MASK) {
-        modifiers |= Modifiers::CONTROL;
-    }
-    if state.contains(gtk4::gdk::ModifierType::ALT_MASK) {
-        modifiers |= Modifiers::ALT;
-    }
-    if state.contains(gtk4::gdk::ModifierType::SUPER_MASK) {
-        modifiers |= Modifiers::META;
-    }
-    let name = key.name()?;
-    let key_name = match name.as_str() {
-        "ISO_Left_Tab" => {
-            modifiers |= Modifiers::SHIFT;
-            "Tab".to_string()
-        }
-        "Return" | "KP_Enter" => "Enter".to_string(),
-        "BackSpace" => "Backspace".to_string(),
-        "Left" => "ArrowLeft".to_string(),
-        "Right" => "ArrowRight".to_string(),
-        "Up" => "ArrowUp".to_string(),
-        "Down" => "ArrowDown".to_string(),
-        "space" => " ".to_string(),
-        _ => match key.to_unicode() {
-            Some(c) if !c.is_control() => c.to_string(),
-            _ => name.to_string(),
-        },
-    };
-    Some(TabCommand::KeyDown {
-        code: key_name.clone(),
-        key: key_name,
-        modifiers,
-    })
-}
-
-fn contains_ci(haystack: &str, needle: &str) -> bool {
-    haystack
-        .as_bytes()
-        .windows(needle.len())
-        .any(|w| w.eq_ignore_ascii_case(needle.as_bytes()))
-}
-
-fn ends_with_ci(s: &str, suffix: &str) -> bool {
-    s.len() >= suffix.len() && s.as_bytes()[s.len() - suffix.len()..].eq_ignore_ascii_case(suffix.as_bytes())
-}
-
-/// Whether to render dark: `GOSUB_COLOR_SCHEME` if set, else the desktop's preference.
-fn desktop_prefers_dark() -> bool {
-    if let Ok(v) = std::env::var("GOSUB_COLOR_SCHEME") {
-        return v.eq_ignore_ascii_case("dark");
-    }
-    // GNOME / freedesktop: org.gnome.desktop.interface color-scheme = 'prefer-dark'.
-    let gnome_dark = gtk4::gio::SettingsSchemaSource::default()
-        .and_then(|src| src.lookup("org.gnome.desktop.interface", true))
-        .filter(|schema| schema.has_key("color-scheme"))
-        .map(|_| gtk4::gio::Settings::new("org.gnome.desktop.interface").string("color-scheme"))
-        .is_some_and(|s| s == "prefer-dark");
-    if gnome_dark {
-        return true;
-    }
-    let Some(settings) = gtk4::Settings::default() else {
-        return false;
-    };
-    settings.is_gtk_application_prefer_dark_theme()
-        || settings
-            .gtk_theme_name()
-            .is_some_and(|n| contains_ci(n.as_str(), "dark"))
-        || std::env::var("GTK_THEME").is_ok_and(|t| ends_with_ci(&t, ":dark"))
-}
-
 fn main() {
     eprintln!(
         "{} v{} — GTK4 browser window, Skia GPU (OpenGL) rendering",
@@ -218,9 +101,8 @@ fn main() {
 
         let backend = SkiaBackend::new();
         let mut engine = GosubEngine::<AppConfig>::new(None, Arc::new(backend), compositor.clone());
-        // Colour scheme: GOSUB_COLOR_SCHEME=dark|light wins; otherwise the desktop preference
-        // (GNOME's color-scheme GSetting, a "-dark" GTK theme, or an app-level prefer-dark flag).
-        if desktop_prefers_dark() {
+        // GOSUB_COLOR_SCHEME=dark renders pages and native controls in the dark scheme.
+        if std::env::var("GOSUB_COLOR_SCHEME").is_ok_and(|v| v.eq_ignore_ascii_case("dark")) {
             let _ = engine.settings().set(
                 "renderer.color_scheme",
                 gosub_config::settings::Setting::String("dark".into()),
@@ -230,16 +112,12 @@ fn main() {
         let event_rx = engine.subscribe_events();
 
         let zone_cfg = ZoneConfig::builder().do_not_track(true).build().expect("ZoneConfig");
-        let cookie_store: gosub_engine::cookies::CookieStoreHandle =
-            SqliteCookieStore::new(".pipeline-browser-cookies.db".into())
-                .expect("failed to open cookie store")
-                .into();
         let zone_services = ZoneServices {
             storage: Arc::new(StorageService::new(
-                Arc::new(SqliteLocalStore::new(".pipeline-browser-local.db").expect("local store")),
+                Arc::new(SqliteLocalStore::new(":memory:").expect("local store")),
                 Arc::new(InMemorySessionStore::new()),
             )),
-            cookie_store: Some(cookie_store),
+            cookie_store: None,
             cookie_jar: None,
             partition_policy: PartitionPolicy::None,
             places: None,
@@ -537,29 +415,6 @@ fn main() {
         });
         gl_area.add_controller(click_ctl);
 
-        // Clicking the page gives the area GTK focus so keys reach it; Tab is swallowed so GTK
-        // doesn't move focus to the address bar.
-        let key_ctl = gtk4::EventControllerKey::new();
-        key_ctl.connect_key_pressed({
-            let tab = tab.clone();
-            move |_, key, _code, state| {
-                let Some(cmd) = key_down_command(key, state) else {
-                    return glib::Propagation::Proceed;
-                };
-                let is_tab = matches!(&cmd, TabCommand::KeyDown { key, .. } if key == "Tab");
-                let tab = tab.borrow().clone();
-                TOKIO_RT.spawn(async move {
-                    let _ = tab.send(cmd).await;
-                });
-                if is_tab {
-                    glib::Propagation::Stop
-                } else {
-                    glib::Propagation::Proceed
-                }
-            }
-        });
-        gl_area.add_controller(key_ctl);
-
         // ── Address bar ──────────────────────────────────────────────────────
 
         let address_entry = Entry::new();
@@ -608,17 +463,8 @@ fn main() {
         glib::spawn_future_local({
             let address_entry = address_entry.clone();
             let status_label = status_label.clone();
-            let da = gl_area.clone();
-            let tab = tab.clone();
             async move {
                 while let Some(evt) = ui_rx.recv().await {
-                    if handle_clipboard_event(&evt, &da, &tab) {
-                        continue;
-                    }
-                    if let EngineEvent::CursorChanged { cursor, .. } = evt {
-                        apply_cursor(&da, cursor);
-                        continue;
-                    }
                     match evt {
                         EngineEvent::Navigation {
                             event: NavigationEvent::Finished { url, .. },
